Remove logging test class
and fix tabs in pom.xml

since we can do the same without having that class.

In case we would lik to be able to see the log output in the maven build
it is possible to define a maven profile providing the sl4j simple
jar alongside with some system properties activating our loggers.

In case we would like to be able to see the log during devlopment
we can configure the launch session accordingly with sl4j simple
added to the class path and also the required system properties.