two different deployments depending on branch: master or release
For the 'master' branch we perform snapshot releases.
For the 'release' branch we perform releases into the sonatype/mavencentral stage.

For all other braches, e.g. feature branches there is no deployment.

There is no check for '-SNAPSHOT', neither that they are present on
the master branch for snapshot release, nor that they are not present
on the release branch for the staging repo.