Perform deployment via nexus-staging-maven-plugin
It is not possibly anymore since to deploy with the
standard maven deploy mechanism into nexus staging repos
since the IP adress of the sender changes between several
artifact uploads. This confuses nexus, resulting in having
the artifacts deployed into several staging repositories.

In order to have all artifacts again in one and the same
staging repo we switch to nexus-staging-maven-plugin